+++ app/src/main/java/com/example/finportfolio/entity/Asset.kt
@@ -1,6 +1,6 @@
 package com.example.finportfolio.entity
 
-data class Asset (
+data class Asset(
     val id: Int,
-    val name: String,
+    val name: String
 )
