+++ app/src/main/java/com/example/finportfolio/fragments/assets/AssetFragment.kt
@@ -4,10 +4,8 @@ import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import androidx.fragment.app.Fragment
-import com.example.finportfolio.BaseFragment
 import com.example.finportfolio.databinding.FragmentAssetBinding
-import com.example.finportfolio.entity.Asset
+import com.example.finportfolio.fragments.BaseFragment
 import com.example.finportfolio.repository.AssetRepository
 import com.example.finportfolio.rv.AssetAdapter
 
@@ -17,7 +15,6 @@ class AssetFragment : BaseFragment<FragmentAssetBinding>() {
         inflater: LayoutInflater,
         container: ViewGroup?
     ): FragmentAssetBinding = FragmentAssetBinding.inflate(inflater, container, false)
-
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
