+++ app/src/main/java/com/example/finportfolio/fragments/portfolio/PortfolioFragment.kt
@@ -2,8 +2,8 @@ package com.example.finportfolio.fragments.portfolio
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
-import com.example.finportfolio.BaseFragment
 import com.example.finportfolio.databinding.FragmentPortfolioBinding
+import com.example.finportfolio.fragments.BaseFragment
 
 class PortfolioFragment : BaseFragment<FragmentPortfolioBinding>() {
