+++ app/src/main/res/layout/fragment_setting.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".data.setting.SettingFragment">
+    tools:context=".fragments.setting.SettingFragment">
 
     <LinearLayout
         android:layout_width="match_parent"
